// src/services/generated/monitoring/models/notificationResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Dexa Group Monitoring Employee API
 * REST API admin HRD untuk master karyawan, absensi read-only, audit, dan notifikasi.
 * OpenAPI spec version: 1.0.0
 */
import type { NotificationResponseDtoData } from './notificationResponseDtoData';

export interface NotificationResponseDto {
  id: string;
  eventId: string;
  /** @nullable */
  employeeId: string | null;
  title: string;
  message: string;
  /** @nullable */
  data: NotificationResponseDtoData;
  /** @nullable */
  readAt: string | null;
  createdAt: string;
}